Create Reorder snake script
This script is a stopgap before a better solution (improvement to the Reshape widget) is found.
The script will reorder the data in a dataset that has been recorded in Snake mode instead of Raster mode. This makes the data well ordered for the Reshape widget.
The script first generates an array with size 'number of spectra' with ncol numbers in ascending order, then ncol numbers in descending order till and so on.
The scritpt then slices the in_data.X according to the reordering. This reverses the order of the spectra on every even row. 
The dataset may then be propagated to a reshape widget to redraw the map wit the correct coordinates.