Create Curve fitting for single spectrum
This is a script for performing the curve fitting analysis of a single spectrum in the Python script widget.
It uses the lmfit package
It doesn't manage the presence of NAN in the spectral data (in_data.X)
It uses AnyQt to query inputs from the user to set initial peak parameters: number of peaks, type of peak profile (gaussian, lorentzian, voigt), amplitude of the peak, half-width of the peaks. The boundaries for the parameters are hard coded in the script.
It performs a first fit and displays a figure where you can check if the fit is satisfactory. If not happy yo ucan rerun the fit, it will randomly wiggle the initial parameters to get a new result. When complete and satisfactory, it will extract and print the peak position, area and relative area (%) of each peak in the console. 
There are a few checks like the peak position should be comprised within the x axis range.